refactor(structure): restructure to full stack with backend/frontend split (closes #8)
