--- apps/insurance-claim-ui/src/pages/ClaimStatusPage.tsx
@@ -1,9 +1,0 @@
-import ClaimStatus from "../components/ClaimStatus"
-
-export default function ClaimStatusPage() {
-  return (
-    <div className="container">
-      <ClaimStatus />
-    </div>
-  )
-}
