--- apps/insurance-claim-ui/src/pages/SubmitClaim.tsx
@@ -1,9 +1,0 @@
-import ClaimForm from "../components/ClaimForm"
-
-export default function SubmitClaim() {
-  return (
-    <div className="container">
-      <ClaimForm />
-    </div>
-  )
-}
