--- apps/insurance-claim-ui/src/styles/main.css
@@ -1,214 +1,0 @@
-* {
-  box-sizing: border-box;
-  font-family: "Segoe UI", Arial, sans-serif;
-}
-
-body {
-  margin: 0;
-  background-color: #f4f6f9;
-}
-
-/* ================= NAVBAR ================= */
-.navbar {
-  background: #0d47a1;
-  color: white;
-  padding: 18px 40px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.logo {
-  font-size: 22px;
-  font-weight: 600;
-}
-
-.nav-links span {
-  font-size: 16px;
-  cursor: pointer;
-  padding-bottom: 4px;
-}
-
-.nav-links .active {
-  border-bottom: 3px solid #ff9800;
-}
-
-/* ================= PAGE LAYOUT ================= */
-.container {
-  padding: 40px;
-  display: flex;
-  justify-content: center;
-}
-
-/* ================= CARD ================= */
-.card {
-  background: #ffffff;
-  width: 100%;
-  max-width: 900px;
-  padding: 30px 35px;
-  border-radius: 12px;
-  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.08);
-}
-
-.card h3 {
-  margin-top: 0;
-  margin-bottom: 10px;
-}
-
-.subtitle {
-  color: #555;
-  margin-bottom: 25px;
-}
-
-/* ================= FORM ================= */
-form label {
-  font-weight: 600;
-  margin-bottom: 6px;
-  display: block;
-}
-
-form input,
-form select,
-form textarea {
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 18px;
-  border-radius: 6px;
-  border: 1px solid #ccc;
-  font-size: 14px;
-}
-
-textarea {
-  resize: vertical;
-  min-height: 90px;
-}
-
-/* Grid layout */
-.row {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 20px;
-}
-
-/* File input */
-input[type="file"] {
-  padding: 6px;
-}
-
-/* Submit button */
-button {
-  background: #1976d2;
-  color: white;
-  border: none;
-  padding: 12px 20px;
-  font-size: 15px;
-  border-radius: 6px;
-  cursor: pointer;
-}
-
-button:hover {
-  background: #0d47a1;
-}
-
-/* ================= STATUS PAGE ================= */
-.success {
-  color: #2e7d32;
-  font-weight: 600;
-  margin-bottom: 10px;
-}
-
-.timeline {
-  list-style: none;
-  padding-left: 0;
-  margin-top: 20px;
-  margin-bottom: 20px;
-}
-
-.timeline li {
-  margin-bottom: 10px;
-  font-size: 15px;
-}
-
-.timeline .danger {
-  color: #d32f2f;
-  font-weight: 600;
-}
-
-/* ================= DETAILS BOX ================= */
-.details {
-  background: #f5f7fb;
-  padding: 15px;
-  border-radius: 8px;
-}
-
-.details p {
-  margin: 8px 0;
-}
-
-.details .danger {
-  color: #d32f2f;
-  font-weight: bold;
-}
-
-/* ================= RESPONSIVE ================= */
-@media (max-width: 768px) {
-  .row {
-    grid-template-columns: 1fr;
-  }
-}
-/* ===== Progress Bar ===== */
-.progress-container {
-  display: flex;
-  justify-content: space-between;
-  margin: 30px 0;
-  position: relative;
-}
-
-.progress-container::before {
-  content: "";
-  position: absolute;
-  top: 18px;
-  left: 0;
-  width: 100%;
-  height: 4px;
-  background: #e0e0e0;
-  z-index: 0;
-}
-
-.step {
-  text-align: center;
-  position: relative;
-  z-index: 1;
-  width: 100%;
-}
-
-.step span {
-  display: block;
-  margin-top: 8px;
-  font-size: 13px;
-}
-
-.circle {
-  width: 18px;
-  height: 18px;
-  background: #e0e0e0;
-  border-radius: 50%;
-  margin: 0 auto;
-}
-
-/* Completed steps */
-.step.completed .circle {
-  background: #2e7d32;
-}
-
-/* Active step */
-.step.active .circle {
-  background: #1976d2;
-}
-
-/* Active label */
-.step.active span {
-  font-weight: 600;
-  color: #1976d2;
-}
-
