+++ apps/smart-claims-processor/backend/data/submitted_claims.csv
@@ -1,0 +1,6 @@
+Timestamp,Customer_ID,Description,Amount,Tenure,Category,Urgency,Fraud_Risk,Anomaly_Score,Status
+2026-01-12 18:09:29,SAFE-001,I went for my routine dental checkup and the doctor suggested a cleaning procedure. Attached is the invoice.,100.0,5,Medical,Medium,Low,0.039066737835938614,Pending Review
+2026-01-12 18:11:00,RISK-999,I lost my expensive watch and some other items while traveling. Not sure exactly where it happened.,200.0,1,Medical,Medium,High,-0.009722439333052613,Pending Review
+2026-01-12 18:12:48,100,cancer stage 4,100.0,8,Medical,Medium,Low,0.03910101881349992,Pending Review
+2026-01-12 18:23:39,100,cancer third stage,100.0,5,22,Medium,Low,0.08622932695359725,Pending Review
+2026-01-12 18:31:38,100,cancer third stage,100.0,1,22,Medium,Low,0.012160848203240593,Pending Review
