+++ apps/smart-claims-processor/backend/requirements.txt
@@ -1,0 +1,6 @@
+pandas
+numpy
+scikit-learn
+streamlit
+faker
+joblib
