+++ apps/smart-claims-processor/frontend/src/components/Navbar.tsx
@@ -1,0 +1,10 @@
+export default function Navbar() {
+  return (
+    <nav className="navbar">
+      <div className="logo">InsuranceClaimFusion AI</div>
+      <div className="nav-links">
+        <span className="active">Claims</span>
+      </div>
+    </nav>
+  )
+}
