+++ apps/smart-claims-processor/frontend/src/pages/ClaimStatusPage.tsx
@@ -1,0 +1,9 @@
+import ClaimStatus from "../components/ClaimStatus"
+
+export default function ClaimStatusPage() {
+  return (
+    <div className="container">
+      <ClaimStatus />
+    </div>
+  )
+}
