+++ apps/smart-claims-processor/frontend/src/pages/SubmitClaim.tsx
@@ -1,0 +1,9 @@
+import ClaimForm from "../components/ClaimForm"
+
+export default function SubmitClaim() {
+  return (
+    <div className="container">
+      <ClaimForm />
+    </div>
+  )
+}
