--- apps/smart-claims-processor/requirements.txt
@@ -1,6 +1,0 @@
-pandas
-numpy
-scikit-learn
-streamlit
-faker
-joblib
